split cluster deletion into multiple functions for readability
